+++ Makefile
@@ -33,7 +33,7 @@ lint_imports:
 	test -z $(goimports -e -d . | tee /dev/stderr)
 
 lint_cyclo:
-	gocyclo -over 40 .
+	gocyclo -over 45 .
 
 lint_golint:
 	golint -set_exit_status $(go list -tags ci ./...)
