+++ pkg/util/util.go
@@ -1,0 +1,20 @@
+package util
+
+import (
+	"io"
+	"log"
+	"os"
+)
+
+func Exists(filePath string) bool {
+	if _, err := os.Stat(filePath); os.IsNotExist(err) {
+		return false
+	}
+	return true
+}
+
+func LogClose(closer io.Closer) {
+	if errClose := closer.Close(); errClose != nil {
+		log.Printf("Error trying to close: %v\n", errClose)
+	}
+}
